vera-ki-api (09.02.26, 16.28)/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
requests>=2.32.0
python-dotenv>=1.0.0
